adjust code
