+++ src/azure-cli-core/azure/cli/core/aaz/_arg_action.py
@@ -483,112 +483,6 @@ class AAZListArgAction(AAZCompoundTypeArgAction):
         raise AAZInvalidValueError(f"list type value expected, got '{data}'({type(data)})")
 
 
-class AAZActionGroupReceiverArgAction(AAZCompoundTypeArgAction):
-
-    def __call__(self, parser, namespace, values, option_string=None):
-        if not isinstance(getattr(namespace, self.dest), AAZArgActionOperations):
-            # overwrite existing namespace value which is not an instance of AAZArgActionOperations.
-            # especially the default value of argument.
-            setattr(namespace, self.dest, AAZArgActionOperations())
-        dest_ops = getattr(namespace, self.dest)
-        try:
-            if self._schema.singular_options and option_string in self._schema.singular_options:
-                # if singular option is used then parsed values by element action
-                action = self._schema.Element._build_cmd_action()
-                action.setup_operations(dest_ops, values, prefix_keys=[_ELEMENT_APPEND_KEY])
-            else:
-                self.setup_operations(dest_ops, values)
-        except (ValueError, KeyError) as ex:
-            raise azclierror.InvalidArgumentValueError(f"Failed to parse '{option_string}' argument: {ex}") from ex
-        except AAZShowHelp as aaz_help:
-            aaz_help.keys = (option_string, *aaz_help.keys)
-            self.show_aaz_help(parser, aaz_help)
-
-    @classmethod
-    def setup_operations(cls, dest_ops, values, prefix_keys=None):
-        if prefix_keys is None:
-            prefix_keys = []
-        if values is None or values == []:
-            if cls._schema._blank == AAZUndefined:
-                raise AAZInvalidValueError("argument cannot be blank")
-            assert not isinstance(cls._schema._blank, AAZPromptInput), "Prompt input is not supported in List args."
-            dest_ops.add(copy.deepcopy(cls._schema._blank), *prefix_keys)
-        else:
-            assert isinstance(values, list)
-            ops = []
-            try:
-                # Standard Expression
-                for key, key_parts, value in cls.decode_values(values):
-                    schema = cls._schema
-                    for k in key_parts:
-                        schema = schema[k]  # pylint: disable=unsubscriptable-object
-                    action = schema._build_cmd_action()
-                    data = action.format_data(value)
-                    ops.append((key_parts, data))
-            except AAZShowHelp as aaz_help:
-                raise aaz_help
-            except Exception as ex:  # pylint: disable=broad-except
-                # This part of logic is to support Separate Elements Expression which is widely used in current command,
-                # such as:
-                #       --args val1 val2 val3.
-                # The standard expression of it should be:
-                #       --args [val1,val2,val3]
-
-                for value in values:
-                    key, _, _ = cls._str_parser.split_partial_value(value)
-                    if key is not None:
-                        # key should always be None
-                        raise ex
-
-                element_action = cls._schema.Element._build_cmd_action()
-                if not issubclass(element_action, AAZSimpleTypeArgAction):
-                    # Separate Elements Expression only supported for simple type element array
-                    raise ex
-
-                # dest_ops
-                try:
-                    element_ops = AAZArgActionOperations()
-                    for value in values:
-                        element_action.setup_operations(element_ops, value, prefix_keys=[_ELEMENT_APPEND_KEY])
-                except AAZShowHelp as aaz_help:
-                    aaz_help.schema = cls._schema
-                    aaz_help.keys = (0, *aaz_help.keys)
-                    raise aaz_help
-
-                elements = []
-                for _, data in element_ops._ops:
-                    elements.append(data)
-                ops = [([], elements)]
-
-            for key_parts, data in ops:
-                dest_ops.add(data, *prefix_keys, *key_parts)
-
-    @classmethod
-    def format_data(cls, data):
-        if data == AAZBlankArgValue:
-            if cls._schema._blank == AAZUndefined:
-                raise AAZInvalidValueError("argument value cannot be blank")
-            assert not isinstance(cls._schema._blank, AAZPromptInput), "Prompt input is not supported in List args."
-            data = copy.deepcopy(cls._schema._blank)
-
-        if data is None:
-            if cls._schema._nullable:
-                return data
-            raise AAZInvalidValueError("field is not nullable")
-
-        if isinstance(data, list):
-            result = []
-            action = cls._schema.Element._build_cmd_action()
-            for idx, value in enumerate(data):
-                try:
-                    result.append(action.format_data(value))
-                except AAZInvalidValueError as ex:
-                    raise AAZInvalidValueError(f"Invalid at [{idx}] : {ex}") from ex
-            return result
-
-        raise AAZInvalidValueError(f"list type value expected, got '{data}'({type(data)})")
-
-
 class AAZGenericUpdateAction(Action):  # pylint: disable=too-few-public-methods
     DEST = 'generic_update_args'
     ACTION_NAME = None
